DXIL debugger discard should not terminate helper lanes
Active lane is demoted to helper invocation which for pixel debug terminates the debug.

Non-active lanes skip over a degenerate branch if that is the next instruction after the discard.

This is a quick fix until maximal re-convergence style control flow is implemented, where quads reconverge at local uniform block